scripts/run_pipeline.py
#!/usr/bin/env python
"""
Example script to run a full data processing and model training pipeline.
"""

import hydra
from omegaconf import DictConfig

@hydra.main(config_path="../config", config_name="config")
def main(cfg: DictConfig) -> None:
    """Main pipeline function."""
    print("Starting pipeline...")
    print("Configuration:")
    print(cfg)

    # 1. Load data
    print("\nStep 1: Loading data...")
    # data = load_data(cfg.data.raw_data_path)

    # 2. Preprocess data
    print("\nStep 2: Preprocessing data...")
    # processed_data = preprocess_data(data)

    # 3. Train model
    print("\nStep 3: Training model...")
    # model = train_model(processed_data, cfg.model, cfg.training)

    # 4. Evaluate model
    print("\nStep 4: Evaluating model...")
    # metrics = evaluate_model(model, processed_data)
    # print(f"Evaluation metrics: {metrics}")

    print("\nPipeline finished.")

if __name__ == "__main__":
    main()
